+++ neoforge/src/main/java/net/spell_engine/neoforge/PlatformImpl.java
@@ -1,11 +1,8 @@
 package net.spell_engine.neoforge;
 
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.ItemStack;
 import net.neoforged.fml.ModList;
 import net.spell_engine.Platform;
-import net.spell_engine.neoforge.compat.accessories.AccessoriesCompat;
-import net.spell_engine.neoforge.compat.accessories.NeoForgeCompatFeatures;
+import net.spell_engine.compat.accessories.NeoForgeCompatFeatures;
 
 public class PlatformImpl {
     public static Platform.Type getPlatformType() {
@@ -21,14 +18,6 @@ public class PlatformImpl {
         @Override
         public void awakeSlotModCompat() {
             NeoForgeCompatFeatures.initSlotCompat();
-        }
-
-        @Override
-        public ItemStack getSpellBookSlot(PlayerEntity player) {
-            if (AccessoriesCompat.isEnabled()) {
-                return AccessoriesCompat.getSpellBookStack(player);
-            }
-            return ItemStack.EMPTY;
         }
     }
     private static final Platform.Util UTIL = new NeoForgeUtil();